backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
pyotp==2.9.0
qrcode[pil]==7.4.2
PyJWT==2.12.1
bcrypt==4.2.1
pyyaml==6.0.2
cryptography==46.0.7
python-multipart==0.0.26
pydantic-settings==2.4.0
alembic==1.13.3
httpx==0.27.2
slowapi==0.1.9
